Part 1 and 2 done. 17 needing review

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: nilim <nilim@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/29 09:28:19 by nilim             #+#    #+#             */
-/*   Updated: 2026/07/30 12:18:47 by nilim            ###   ########.fr       */
+/*   Updated: 2026/08/05 09:52:11 by nilim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	size_t			i;
+	int				i;
 	unsigned char	*tgt;
 
 	i = 0;
