+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: nilim <nilim@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/08/01 11:16:09 by nilim             #+#    #+#             */
-/*   Updated: 2026/08/04 12:55:36 by nilim            ###   ########.fr       */
+/*   Updated: 2026/08/05 10:03:16 by nilim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 char	*ft_strchr(const char *str, int c)
 {
-	unsigned int	i;
+	int	i;
 
 	i = 0;
-	while (i <= (unsigned int)ft_strlen(str))
+	while (i <= ft_strlen(str))
 	{
 		if (str[i] == c)
 			return ((char *)&(str[i]));
