+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: nilim <nilim@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/08/01 12:01:10 by nilim             #+#    #+#             */
-/*   Updated: 2026/08/04 12:56:09 by nilim            ###   ########.fr       */
+/*   Updated: 2026/08/05 10:11:08 by nilim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 char	*ft_strrchr(const char *str, int c)
 {
-	size_t	n;
-	size_t	i;
+	int	n;
+	int	i;
 
 	n = ft_strlen(str);
 	i = 0;
-	while (i < (size_t)ft_strlen(str))
+	while (i < ft_strlen(str))
 	{
 		if (str[n] == c)
 			return ((char *)&(str[n]));
